--- src/store/unauthorized.store.ts
@@ -1,13 +1,0 @@
-'use client'
-
-import { create } from 'zustand'
-
-type UnauthorizedStore = {
-    showModal: boolean
-    setShowModal: (value: boolean) => void
-}
-
-export const useUnauthorizedModalStore = create<UnauthorizedStore>((set) => ({
-    showModal: false,
-    setShowModal: (value) => set({ showModal: value }),
-}))
